feat: players services from firebase
